refactor tests (split files)
