+++ trusted-verifier/tests/integration/mod.rs
@@ -3,5 +3,7 @@
 //! This module contains integration tests that require external services
 //! (such as running Aptos chains).
 
-pub mod aptos_client_test;
+pub mod connectivity_test;
+pub mod deployment_test;
+pub mod event_polling_test;
